Small fixes to avoid compiler warnings and provide a warranty that template functions like 'loadFromFile' are instantiated in karabo (libkarabo.so) but karathon (karathon.so) and others (for example, libiibus.so).
git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/trunk@12804 a23ab9f9-1ea2-4934-960d-990259f9ea8c

+++ src/karabo/util/TimeProfiler.cc
@@ -234,7 +234,7 @@ namespace karabo {
             vector<string> keys;
             period.getKeys(keys);
 
-            for (int i = 1; i < keys.size(); ++i) {
+            for (size_t i = 1; i < keys.size(); ++i) {
                 if ((keys[i] == "KRB_start") || (keys[i] == "KRB_stop") || (keys[i] == "KRB_duration") || (keys[i] == "KRB_details")) continue;
                 sql(os, keys[i], period.get<Hash>(keys[i]), current_key);
             }
